Partially revert self-upvote UI bug fix.
Turns out the snippet in templates/comments.html was necessary to
get proper behavior in notifications.

Might come back to debug this later, but people use notifications more
than they deal with edge case self-upvotes, so reverting for now.

The changes to prevent coin fuckery with self-upvoting appear to work
correctly in the wild, so leaving those in place.